Saner delta.Command.apply signature
The `Command.apply()` method currently returns a `(schema, object)`
tuple.  The reason is historical: originally the method returned just
the object, but with Schema immutability refactoring it was adjusted to
also returned the schema.  Now, there's no real reason why `apply()`
needs to return the object, since it can easily be accessed as the
`cmd.scls` attribute.  While at it, fix the occasional `context=None`
signatures that are also a historical relic.